Move type definitions to `_types.py`
